requirements.txt
aiogram>=3.4,<4
aiohttp>=3.9,<4
beautifulsoup4>=4.12,<5
lxml
python-dotenv
